[#19] Add flask app mocking the EC API
This is a small flask app that can be run while developing or as part of
the tests to mock the EC API. It performs basic Authorization and
Validation checks that mimic the spec provided by Microsoft so far.

The responses are also in accordance to the MS spec.

Current supported operations are Request dataset creation and Request
dataset update.