[#14] Modify package_create behaviour and add auth tests
We want all dataset creations to be forwarded to the EC API, so we
override the default `package_create` forwarding requests to
`dataset_request_create`.

To still have the ability to create datasets locally (and not request
their creation to EC), an `local_action` option has been added to the
context. Only sysadmins will be able to do it. This will be useful eg to
create harvest sources.

Auth tests has been added but settings like whether users need to belong
to organizations to create a dataset or allowing unowned datasets need
to be discussed.