[#14] Add custom landing page for pending datasets
If a dataset has been submitted for creation to the EC API, the dataset
page will show a temporary page until the request has been completed. It
will probably ping the server periodically to check the status and
refresh when ready. This is left for the next iteration.

On most cases dataset creation should take seconds so perhaps we can
bypass this page entirely.