[#16] Add action for requesting file creation
Similarly to `dataset_request_create`, `file_request_create` will validate
a resource dict, ping the EC API and store the request id in
`task_status` (in fact the whole thing needs to be refactored as there
is quite a lot of duplicated code).

TODO: this does not deal with the actual data / file contents yet.