[#45] Match Platform and CKAN ids
From the latest iteration ids for objects in the platform are no longer
integers but GUIDs. We will use the same value for CKAN ids to make
operations easier. Most objects in CKAN allow you to set the id, except
datasets. In this case though, we allow it via our custom schema.

Note that datasets ids are still integers in the platform, but this will
change in next iteration.