[#56] Store last harvested AuditId
The way the changelog API works, you need to provide a starting AuditId
from which the more recent ones will be returned. We need to persist
this somehow, so we've added a DB table to store it with a timestamp.

The provided AuditId is returned as well, so we need to skip the first
item.

DB tables are created on startup if not already there.