[#56] Update file creation handling
From iteration 4 all audits related to File creation have a VersionId
regardless of whether they are new files or new file versions. We need
to always call the file_vesion_show endpoint and decide if we need to
create or updated by checking if the resource exists or not.

Note that the necessary endpoint is down now on the platform so this can
not be live tested.