[#90] Implement file updates
We have used the Update File endpoint (3.1.2.11 on lastest spec) rather
than the new file version as this one allows not resending the uploaded
file.

We are removing the url if it looks like the original uploaded file.

We also need to store the version id against the resource, as it is
required on the URL used.